+++ src/date.rs
@@ -121,3 +121,18 @@ mod date_parsing_tests {
     }
 }
 
+pub fn unique_year_month(start_date: Date, end_date: Date) -> (Option<u16>, Option<u8>) {
+    let y1 = start_date.year;
+    let y2 = end_date.year;
+    if y1 == y2 {
+        let m1 = start_date.month;
+        let m2 = end_date.month;
+        if m1 == m2 {
+            (Some(y1), Some(m1))
+        } else {
+            (Some(y1), None)
+        }
+    } else {
+        (None, None)
+    }
+}