+++ src/md/printer.rs
@@ -2,6 +2,7 @@ use {
     super::*,
     crate::*,
     have::Fun,
+    itertools::*,
     minimad::{OwningTemplateExpander, TextTemplate},
     num_format::{Locale, ToFormattedString, WriteFormatted},
     std::{
@@ -64,8 +65,7 @@ impl Printer {
         let color = args.color.value().unwrap_or(!is_output_piped());
         let skin = skin::make_skin(color);
         let key = args.key;
-        let date_filter = args.date.as_ref()
-            .and_then(|p| log_base.make_date_filter(p).ok());
+        let date_filter = log_base.filterer.date_filter().map(|f| f.clone());
         let changes = args.changes;
         let all_paths = args.all;
         Self {