+++ src/md/summary.rs
@@ -5,20 +5,38 @@ use {
 };
 
 static SUMMARY_MD: &str = r#"
- ${hits} hits and ${bytes} from **${start}** to **${end}**
+${hits} hits and ${bytes} from **${start}** to **${end}**
+${filterings
+Filtering by ${field} on pattern `${pattern}` removed **${removed_percent}** of total lines
+}
+${filtered-stats
+ **==>** ${hits} hits and ${bytes}
+}
 "#;
 
-pub fn print_summary(log_base: &LogBase, printer: &Printer) {
+pub fn print_summary(base: &LogBase, printer: &Printer) {
     let mut expander = OwningTemplateExpander::new();
-    let bytes = log_base.lines
-        .iter()
-        .map(|line| line.bytes_sent)
-        .sum();
+    let total_bytes = base.unfiltered_histogram.total_bytes_sent();
     expander
-        .set_md("hits", printer.md_hits(log_base.lines.len()))
-        .set_md("bytes", printer.md_bytes(bytes))
-        .set("start", log_base.start_time())
-        .set("end", log_base.end_time());
+        .set_md("hits", printer.md_hits(base.unfiltered_count as usize))
+        .set_md("bytes", printer.md_bytes(total_bytes))
+        .set("start", base.start_time())
+        .set("end", base.end_time());
+    if base.filterer.has_filters() {
+        let total_hits = base.unfiltered_count as f32;
+        for filtering in &base.filterer.filterings {
+            let removed = filtering.removed_count as f32;
+            let percent = format!("{:.2}%", 100f32 * removed / total_hits);
+            expander.sub("filterings")
+                .set("field", filtering.filter.field_name())
+                .set("pattern", &filtering.pattern)
+                .set("removed_percent", percent);
+        }
+        let filtered_bytes = base.filtered_histogram.total_bytes_sent();
+        expander.sub("filtered-stats")
+            .set_md("hits", printer.md_hits(base.filtered_count as usize))
+            .set_md("bytes", printer.md_bytes(filtered_bytes));
+    }
     printer.print(expander, SUMMARY_MD);
 }